+++ src/store/index.ts
@@ -1,3 +1,5 @@
-export * from './datamodel/DataModel';
+export * from './datamodel';
+export * from './hooks';
 export * from './StoreProvider';
+
 export type { Store } from './Store';
